Extract edit op import helpers
